UBUNTU: SAUCE: apparmor5.1.0 [17/25]: apparmor: mqueue: add fine grained mediation of posix mqueues
BugLink: http://bugs.launchpad.net/bugs/2028253

Add fine grained mediation of posix mqueues. Specifically this patch
adds support for differentiating mqueues based on the name in the ipc
namespace. A follow on patch will add support for implied labels, and
a third patch explicit labels. This is done in part because of
dependencies on other patches to apparmor core.

Signed-off-by: John Johansen <john.johansen@canonical.com>
Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>